SLIP module clean-up
The internal APIs of the SLIP module had some type inconsistencies and some
comments were mangled. This commit cleans up the module and also adds some
error checking that was missing.

Also, variable length arrays (VLAs) were turned into fixed size arrays, because
MSVC does not support VLAs.

+++ lib/platform/posix/serial_posix.c
@@ -179,7 +179,7 @@ int Serial_read(unsigned char * c, unsigned int timeout_ms)
     }
 }
 
-int Serial_write(const unsigned char * buffer, unsigned int buffer_size)
+int Serial_write(const unsigned char * buffer, size_t buffer_size)
 {
     if (fd < 0)
     {
